+++ src/main/java/com/epam/reportportal/utils/TestCaseIdUtils.java
@@ -16,15 +16,23 @@
 
 package com.epam.reportportal.utils;
 
+import com.epam.reportportal.annotations.TestCaseId;
 import com.epam.reportportal.annotations.TestCaseIdKey;
 import com.epam.reportportal.service.item.TestCaseIdEntry;
-import io.reactivex.annotations.Nullable;
 
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.function.Function;
+import java.util.stream.Collectors;
+
+import static java.util.Optional.ofNullable;
 
 /**
- * @author <a href="mailto:ivan_budayeu@epam.com">Ivan Budayeu</a>
+ *
  */
 public class TestCaseIdUtils {
 
@@ -32,26 +40,73 @@ public class TestCaseIdUtils {
 		//static only
 	}
 
+	private static final Function<List<Object>, String> TRANSFORM_PARAMETERS = it -> it.stream()
+			.map(String::valueOf)
+			.collect(Collectors.joining(",", "[", "]"));
+
+	public static String getCodeRef(@Nonnull final Method method) {
+		return method.getDeclaringClass().getCanonicalName() + "." + method.getName();
+	}
+
 	@Nullable
-	public static TestCaseIdEntry getParameterizedTestCaseId(Method method, Object... parameters) {
+	public static String getParametersForTestCaseId(Method method, List<Object> parameters) {
+		if (method == null || parameters == null || parameters.isEmpty()) {
+			return null;
+		}
 		Annotation[][] parameterAnnotations = method.getParameterAnnotations();
+		List<Integer> keys = new ArrayList<>();
 		for (int paramIndex = 0; paramIndex < parameterAnnotations.length; paramIndex++) {
 			for (int annotationIndex = 0; annotationIndex < parameterAnnotations[paramIndex].length; annotationIndex++) {
 				Annotation testCaseIdAnnotation = parameterAnnotations[paramIndex][annotationIndex];
 				if (testCaseIdAnnotation.annotationType() == TestCaseIdKey.class) {
-					return getTestCaseId((TestCaseIdKey) testCaseIdAnnotation, paramIndex, parameters);
+					keys.add(paramIndex);
 				}
 			}
 		}
-		return null;
+		if (keys.isEmpty()) {
+			return TRANSFORM_PARAMETERS.apply(parameters);
+		}
+		if (keys.size() <= 1) {
+			return String.valueOf(parameters.get(keys.get(0)));
+		}
+		return TRANSFORM_PARAMETERS.apply(keys.stream().map(parameters::get).collect(Collectors.toList()));
 	}
 
-	@Nullable
-	private static TestCaseIdEntry getTestCaseId(TestCaseIdKey testCaseIdKey, int paramIndex, Object... parameters) {
-		Object testCaseIdParam = parameters[paramIndex];
-		if (testCaseIdParam != null) {
-			return new TestCaseIdEntry(String.valueOf(testCaseIdParam));
+	public static TestCaseIdEntry getTestCaseId(@Nullable TestCaseId annotation, @Nullable Method method,
+			@Nullable List<Object> parameters) {
+		if (annotation != null) {
+			if (annotation.value().isEmpty()) {
+				if (annotation.parametrized()) {
+					return ofNullable(getParametersForTestCaseId(method, parameters)).map(TestCaseIdEntry::new)
+							.orElse(getTestCaseId(method, parameters));
+				} else {
+					return getTestCaseId(method, parameters);
+				}
+			} else {
+				if (annotation.parametrized()) {
+					return ofNullable(getParametersForTestCaseId(method, parameters)).map(p -> new TestCaseIdEntry(
+							annotation.value() + (p.startsWith("[") ? p : "[" + p + "]"))).orElse(getTestCaseId(method, parameters));
+				} else {
+					return new TestCaseIdEntry(annotation.value());
+				}
+			}
 		}
-		return null;
+		return getTestCaseId(method, parameters);
+	}
+
+	public static TestCaseIdEntry getTestCaseId(@Nullable Method method, @Nullable List<Object> parameters) {
+		return ofNullable(method).map(m -> getTestCaseId(getCodeRef(m), parameters)).orElse(getTestCaseId(parameters));
+	}
+
+	public static TestCaseIdEntry getTestCaseId(@Nullable String codeRef, @Nullable List<Object> parameters) {
+		return ofNullable(codeRef).map(r -> new TestCaseIdEntry(codeRef + ofNullable(parameters).map(TRANSFORM_PARAMETERS).orElse("")))
+				.orElse(getTestCaseId(parameters));
+	}
+
+	public static TestCaseIdEntry getTestCaseId(@Nullable List<Object> parameters) {
+		if (parameters == null || parameters.isEmpty()) {
+			return null;
+		}
+		return new TestCaseIdEntry(TRANSFORM_PARAMETERS.apply(parameters));
 	}
 }