+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Utilities
+  - More Test Case ID methods
+  - CodeRef generation method
+- A proxy parameter handling: `rp.http.proxy=http://localhost:8981`
 
 ## [5.0.2]
 ### Added